INFRA-6741 Fix mime.types assumption
Do not check for file existance and then exit
if it does not exist. This test was added recently.
It is a very difficult thing to test for on
several platforms in different states.

Issue: https://khanacademy.atlassian.net/browse/INFRA-6741

Test plan:

It should run on a clean mac again