Install python2 in cases where there's no system python2
MacOS 12.3 no longer includes a system python 2. The dotfiles
assumed that it did. This instead checks the return code of
`which python2` to see if no python2 is installed at all
(and then installs it in that case)

More detail:
https://khanacademy.slack.com/archives/C02NMB1R5/p1649879303389049?thread_ts=1649877395.083359&cid=C02NMB1R5
and
https://khanacademy.slack.com/archives/C02NMB1R5/p1649090083757979?thread_ts=1648850401.466419&cid=C02NMB1R5

Issue: XXX-XXXX

Test plan:
Jonathan Price ran this on a fresh MacOS 12.3 and it worked.