INFRA-7843 - Move khan-dotfiles to node 16
Jenkins has been upgraded and it is now time to move everybody
forward to node 16.

Note: We install node-16 if there is no version of node inestalled.
However, if there is a version of node installed, we prompt the user
with commands to upgrade node in case they have a reason to stay at
a different version.

Issue: https://khanacademy.atlassian.net/browse/INFRA-7843

Test plan:

* Run `mac-setup-normal.sh` on a system with node 12 installed.
* Run `mac-setup-normal.sh` on a system without node installed.