Prefer x86 version of go on ARM architecture.
Localproxy uses a library, `wasmtime-go`, which is presently does not distribute an ARM architecture go binary, preventing it from running without a number of extra steps.

For now, the easiest thing to do is to always install the ARM architecture version instead.

This also switching the node install to use the same brew86 function, and installs a brew86 alias in the dotfiles if on an ARM architecture Mac.

Test plan:
Uninstall go
Run `make`
Verify the go is installed, and is version `1.16.x darwin/amd64` (not `darwin/arm64`)