Switched info for echo in elevated mac script
I ran into issues run the dotfiles because the dotfile needed to reinstall protobuf, which write a message to console using the `info` function.  The problem is that `info` function is not defined in the elevated script so that failed the rest of the dotfiles installation process.  I just switched out `info` for `echo` and called it a day.  Alternatively I could move all the print helper functions to the shared script but that didnt seem necessary.

Once that was fixed I was able to run dotfiles to completion which ultimately installed `fastly` which was the issue I was trying to fix in the first place.

Issue: "none"

Test plan:
run `make` in the dotfile dir and verify it runs all the way to the end and installs `fastly`.