Install python3.8 as part of khan-dotfiles.
We need this to run python services such as ai-guide-core.  Not
everyone needs this -- not everyone develops on python, and if we use
local-musketeers then the container will already have python3.8 and
the user's local development environment won't need it -- but it's
safest to install it for everyone, and can't hurt anything.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9196

Test plan:
I ran the new commands manually on my linux box, with success.  It's
fingers crossed for the os x changes.