Install python 3.11 on developer machines. (#95)
## Summary:
We don't use it anywhere yet, but when we update ai-guide-core to use
python3.11 -- which we will soon -- we'll be ready!

Note that installing python3.11 does not change any working behavior
of the developer machines; the new python is not on the path, and the
/usr/bin/python3 symlink does not change to point to python3.11.  The
only way to access it is to explicitly run something via
```
python3.11 <something>
```

Nothing does that at the moment, but when we update ai-guide-core its
Makefile rules will start doing exactly that!

Issue: GL-1195

## Test plan:
I ran the new linux-setup.sh commands manually.  Someone will
have to do the same with the os-x commands.

Author: csilvers

Reviewers: ChaseC99, MiguelCastillo, jvoll, PeppeMiller

Required Reviewers:

Approved By: ChaseC99, MiguelCastillo

Checks:

Pull Request URL: https://github.com/Khan/khan-dotfiles/pull/95