Add github.com to the known-hosts file for all users.
This isn't technically required; if we don't do it ssh will just
propmt for permission to add github.com to the known-hosts file when
you first try to connect.

But it's friendly to do it, and if github ever changes their host
keys, then re-running khan-dotfiles would be a way to automatically
fix up your known-hosts file to match.

We used to just do this for some users, but with this PR we now do it
for everybody.

Issue: https://github.com/Khan/khan-dotfiles/pull/93#discussion_r1425626804

Test plan:
I ran the new commands manually, twice, with success.  They do emit a
lot of kinda-scary looking output, I'm wondering if we should try to
suppress some of it.