Fix the linux install based on testing.
It looks like the linux install of khan-dotfiles has bitrotted a bit.
This fixes it up based on a new test regime that occured to me: just
use a docker container for testing!  I wrote up how to do this in the
README.

The main changes I made:
* Add a few packages that might not be installed by default
* Make sure /usr/bin/python is installed
* Make sure the git ssh key is installed correctly

Test plan:
As described in the README.  After running `make` I could run
`make serve` (in webapp) with success!