Force homebrew formulas for icu4c 73.2
Since the node@16 formula is deprecated it will no
longer be updated to work with its latest
dependencies, namely icu4c. The latest is
currently icu4c 74.2 which breaks node@16 which it
is hard coded to look for. Force reinstall of
older icu4c 73.2 formula until we get off of
node@16.

https://github.com/Homebrew/homebrew-core/commit/a16dac3924fa87e5b96e2a1fc98d1fa4ee588c38

postgresql@14 was updated on February 22nd to a
hard coded dependency of icu4c 74.2 so also force
install of an older postgresql@14 formula that
has bindings for 73.2.

https://github.com/Homebrew/homebrew-core/commit/231ce1c9490463ad7b32ef3145a312ee1d6dc968

Issue: https://khanacademy.slack.com/archives/C04SEFXQBNU/p1709169846656299

Test plan:

1. brew services stop postgresql
2. brew uninstall postgresql@14
3. brew uninstall node@16
4. make
5. psql -tc "SELECT rolname from pg_catalog.pg_roles" postgres
6. node -v
7. (in webapp) make serve