Use the homebrew pip3, rather than the system pip3, if it exists.
I saw that setup.sh wanted to use the homebrew python3 to install
keeper, and I lost that when I moved the keeper install to use pip3
instead of `python3 -m pip`.  This PR changes the functionality back
to how it was.

I don't know how much it matters, but it's probably a good practice to
use brew's python for everything.

Issue: none

Test plan:
Fingers crossed.  I don't have a mac to test on.