Don't do ssh-add in bash profile on Linux
Previously, the bash profile would run `ssh-add -K` on all systems.
This caused a problem on Linux, because the `-K` option means
"Load resident keys from a FIDO authenticator" on non-MacOS versions
of OpenSSH. Since the bash profile is sourced by the bashrc file,
any new shell session on Linux prompts the user for an authenticator
PIN instead of showing a prompt.

This commit adds a guard to only run `ssh-add` on MacOS, so shell
sessions won't be blocked on Linux machines.

Issue: none

Test plan:
- Open a terminal emulator on Linux and verify that a prompt appears.