Update to java 21.
We are using java 21 on jenkins now, since that's what Jenkins needs
to run.  So let's use it everywhere!  These days it really only
matters for dataflow, since that's the only thing we have that runs
java.  (I think.  Maybe mobile does for something?)  Which most people
never run.  So it doesn't matter much in any case, but we may as well
let you run dataflow if you have to.

Issue: none

Test plan:
Need to test this on mac.