Install the latest version of protoc.
In webapp's #29166, I updated the protoc libraries -- the things that
emit compiled protobufs in different languages -- to the latest
version.  Now it's time to update the protoc compiler itself.  This is
harder because, sadly, the compiler isn't packaged, and isn't a go
module.  Instead, it's a binary that you have to install.  We do that
in khan-dotfiles.  This PR changes khan-dotfiles to install the latest
version.

To make the transition easier -- and to make future updates to protoc
easier -- I now version the binary and the include directory.  So
running this won't break any current webapp branches.  And when it's
time to update webapp to use the new protoc, I just need to point it
to the new binary name (and new include folders) and things will work.

Issue: none

Test plan:
I ran this manually, and then did:
```
% protoc --version
libprotoc 3.4.0
% protoc-29.3 --version
libprotoc 29.3
```