Add my version of git mm as an alias, but with helpful link so as not to stomp on people (#144)
## Summary:
We really need the `git mm` alias for the merge queue flow. I started w Jeff Yates's version, and made it more robust. It just syncs your local deploy branch with changes from `origin/<master|main>`, but it doesn't error in the case where your _remote_ deploy branch could have been deleted (which happens with merge queues when they merge). And it knows the diff between `master` and `main`.

It seems like this may be a somewhat common alias that people already have, and I wanted to make sure that if we override their version that they have a way to override it back. This pull request adds a sub-command alias with a link that redirects to a question on an FAQ that answers how to do that. We can reuse this for other aliases we want to add that could also step on toes. 

<img width="585" alt="Screenshot 2025-05-08 at 2 50 20 PM" src="https://github.com/user-attachments/assets/a1de559a-2c7c-4a7d-adee-a4e0ec38a919" />

I thought about this problem a lot over the last few months and I think this is a fairly elegant, lowest-touch solution.

Issue: FEI-6635


## Test plan:
- Run `git mm` in a repo that has `.gitconfig.khan` included, see that it runs mine
- Follow link: https://khanacademy.org/r/gitfaq#id-6d95
- Try instructions for putting yours back (if you have one), see that _that_ works

Author: lillialexis

Reviewers: jeresig, csilvers, justdigi, somewhatabstract

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/khan-dotfiles/pull/144