Remove some khan-dotfiles dockerfile stuff.
There was an abortive attempt a while ago to try to build a docker
container that was set up like khan-dotfiles.  But it never worked and
we're not pursuing that approach at the moment, so let's get rid of
the code, which is otherwise confusing.

Issue: none

Test plan:
Fingers crossed